requirements.txt
fastapi
python-multipart
uvicorn
python-docx
PyPDF2
google.generativeai
Pillow
jinja2
